refactoring-unfinished
